[run_prediction] Set dtype to uchar when calling save_image
This fixes certain failure cases with nifti images in which the
conformed always gets saved with a wrong dtype (int32), despite
both the image array and header being explicitly set to uchar.
This leads to the same image failing the conform check at the beginning
of the surf pipeline.

The fix is to ensure that set_data_dtype() is used on the MGHImage object
itself, which is now done in the imported data_utils function.

+++ FastSurferCNN/run_prediction.py
@@ -179,7 +179,8 @@ class RunModelOnData:
             orig_data = np.asanyarray(orig.dataobj)
 
         # Save conformed input image
-        self.save_img(self.subject_conf_name, orig_data, orig)
+        self.save_img(self.subject_conf_name, orig_data, orig, dtype=np.uint8)
+
         return orig, orig_data
 
     def set_subject(self, subject: str, sid: Union[str, None]):
@@ -242,7 +243,7 @@ class RunModelOnData:
         else:
             header = orig.header
 
-        du.save_image(header, orig.affine, np_data, save_as)
+        du.save_image(header, orig.affine, np_data, save_as, dtype=dtype)
         LOGGER.info("Successfully saved image as {}".format(save_as))
 
     def set_up_model_params(self, plane, cfg, ckpt):
